requirements.txt
boto3==1.26.94
botocore==1.29.94
cffi==1.15.1
cryptography==39.0.2
dynamodb-json==1.3
importlib-metadata==6.1.0
jaraco.classes==3.2.3
jmespath==1.0.1
keyring==23.13.1
more-itertools==9.1.0
Pillow==9.4.0
pip==22.3.1
pycparser==2.21
pyirsdk==1.3.3
pypiwin32==223
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
pywin32==305
PyYAML==6.0
s3transfer==0.6.0
setuptools==65.5.1
simplejson==3.18.4
six==1.16.0
sv-ttk==2.4.2
urllib3==1.26.15
wheel==0.38.4
zipp==3.15.0
